Gate AIChatProvider
